+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-#2.1.8
+#2.2.0
+- Implemented new feature 'Move navbar to searchbar'
 - Great performance improvements
 - Favicons now should load faster
 - Project restructured
 - Changed extension name to 'G-Tiles'
 - Added setting 'Change cursor over tile to pointer'
-- Implemented new feature 'Move navbar to searchbar'
+- Fixed issues on settings page
 
 #2.1.7
 - Fixed layout issues for some search queries